packages/auth/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bonanza-auth"
version = "0.2.1"
description = "🔐 Bonanza Labs ✦ Auth — AI agent authentication & identity"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{name = "Bonanza Labs", email = "passiveassets@proton.me"}]
dependencies = ["pydantic>=2.0", "httpx>=0.25", "click>=8.0", "rich>=13.0", "pyjwt>=2.8", "cryptography>=41.0"]

[project.scripts]
bonanza-auth = "bonanza_auth.cli:main"

[project.optional-dependencies]
api = ["fastapi>=0.104", "uvicorn>=0.24"]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]